Decompose RecipeFormBase (877 -> focused units)
Phase 5: split the 877-line RecipeFormBase into a thin composition root
(94 lines) plus single-responsibility units, all under ~300 lines:
- useRecipeForm / useIngredientSections — form + ingredient-section state,
  drag-drop reorder, and the snake_case payload normalizer
- IngredientSectionsEditor (+ SectionAddRow), RecipeMetaFields (+ SpicesEditor)
- types.ts (RecipeFormValues/DragTarget/parseAmount), styles.ts (shared
  Tailwind class strings)
Structural only — rendered DOM/classNames and the wire payload unchanged
(ingredient_sections stays snake_case).

Tests-first: +15 tests, incl. a payload characterization guard asserting
the snake_case key, title strip/auto-title, comma-decimal parsing, and
drag-reorder. 248 tests green; build + lint + check:no-hex clean.
